uniform gaps
With the new gap creating algorithm, every brick shrinks a fixed amount in every direction, no matter what scale of dimensions it has. this results in uniform gaps, in contrast to when you would just rescale the object uniformly: then you get bigger gaps in the direction the part is longer.